% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_fisheries_data.R
\name{fish_summary}
\alias{fish_summary}
\title{Abundance and revenue information of fish caught in Moorea, French Polynesia}
\usage{
fish_summary(catch_location_data, price_data, graph = FALSE)
}
\arguments{
\item{catch_location_data}{data frame with columns northside, westside, and eastside (sides of the island), and rows parrotfish, unicronfish, bonito. mahi, yellowfin, and swordfish (fish species caught)}

\item{price_data}{data frame with column price (Polynesian Franc/kg) and rows parrotfish, unicorn, bonito, mahi, yellowfin, and swordfish (fish species caught)}

\item{graph}{specify TRUE for output of a graph of revenue by location}

\item{total_revenue}{specifiy TRUE for total revenue sum}
}
\value{
returns a list containing  most frequently caught fish in each location, total revenue for each location, and total fisheries revenue sum
}
\description{
calculate the most frequently caught fish in each location, total revenue for each location, total fisheries revenue sum, and if requested a graph of revenue by location and total revenue (as text)
}
